now can print out time of trigger and event packet

+++ SciEvent.cpp
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-void SciEvent::set_mode_(const char* packet_buffer, uint16_t packet_len) {
+void SciEvent::set_mode_(const char* packet_buffer, size_t packet_len) {
 	mode = 0;
 	for (int i = 0; i < 2; i++) {
 		mode <<= 8;
@@ -12,7 +12,7 @@ void SciEvent::set_mode_(const char* packet_buffer, uint16_t packet_len) {
 	mode >>= 7;
 }
 
-void SciEvent::set_ct_num_(const char* packet_buffer, uint16_t packet_len) {
+void SciEvent::set_ct_num_(const char* packet_buffer, size_t packet_len) {
 	ct_num = 0;
 	for (int i = 0; i < 2; i++) {
 		ct_num <<= 8;
@@ -20,7 +20,7 @@ void SciEvent::set_ct_num_(const char* packet_buffer, uint16_t packet_len) {
 	}
 }
 
-void SciEvent::set_timestamp_(const char* packet_buffer, uint16_t packet_len) {
+void SciEvent::set_timestamp_(const char* packet_buffer, size_t packet_len) {
 	timestamp = static_cast<uint8_t>(packet_buffer[17]);
 	for (int i = 0; i < 2; i++) {
 		timestamp <<= 8;
@@ -28,7 +28,7 @@ void SciEvent::set_timestamp_(const char* packet_buffer, uint16_t packet_len) {
 	}
 }
 
-void SciEvent::set_deadtime_(const char* packet_buffer, uint16_t packet_len) {
+void SciEvent::set_deadtime_(const char* packet_buffer, size_t packet_len) {
 	deadtime = 0;
 	for (int i = 0; i < 2; i++) {
 		deadtime <<= 8;
@@ -36,7 +36,7 @@ void SciEvent::set_deadtime_(const char* packet_buffer, uint16_t packet_len) {
 	}
 }
 
-void SciEvent::update(const char* packet_buffer, uint16_t packet_len) {
+void SciEvent::update(const char* packet_buffer, size_t packet_len) {
 	set_mode_(packet_buffer, packet_len);
 	set_ct_num_(packet_buffer, packet_len);
 	set_timestamp_(packet_buffer, packet_len);
