+++ SciTrigger.cpp
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-void SciTrigger::set_mode_(const char* packet_buffer, uint16_t packet_len) {
+void SciTrigger::set_mode_(const char* packet_buffer, size_t packet_len) {
 	mode = 0;
 	for (int i = 0; i < 2; i++) {
 		mode <<= 8;
@@ -10,7 +10,7 @@ void SciTrigger::set_mode_(const char* packet_buffer, uint16_t packet_len) {
 	}
 }
 
-void SciTrigger::set_timestamp_(const char* packet_buffer, uint16_t packet_len) {
+void SciTrigger::set_timestamp_(const char* packet_buffer, size_t packet_len) {
 	timestamp = 0;
 	for (int i = 0; i < 4; i++) {
 		timestamp <<= 8;
@@ -18,7 +18,7 @@ void SciTrigger::set_timestamp_(const char* packet_buffer, uint16_t packet_len) 
 	}
 }
 
-void SciTrigger::set_packet_num_(const char* packet_buffer, uint16_t packet_len) {
+void SciTrigger::set_packet_num_(const char* packet_buffer, size_t packet_len) {
 	packet_num = 0;
 	for (int i = 0; i < 2; i++) {
 		packet_num <<= 8;
@@ -26,7 +26,7 @@ void SciTrigger::set_packet_num_(const char* packet_buffer, uint16_t packet_len)
 	}
 }
 
-void SciTrigger::update(const char* packet_buffer, uint16_t packet_len) {
+void SciTrigger::update(const char* packet_buffer, size_t packet_len) {
 	set_mode_(packet_buffer, packet_len);
 	set_timestamp_(packet_buffer, packet_len);
 	set_packet_num_(packet_buffer, packet_len);
